apps/blob-cron/pyproject.toml
[tool.poetry]
name = "njmtech-blob-cron"
version = "0.0.3"
description = "A cron job to scan Vercel Blob Storage and transform text files to markdown using AI."
authors = ["Gemini <gemini@google.com>"]
readme = "README.md"
packages = [{include = "njm_blob_cron", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.0"
vercel-blob = "^0.4.2"
agno = "0.0.2"
ollama = "^0.2.0"
httpx = "^0.27.0"
asyncpg = "^0.30.0"
infisical-python-sdk = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
